+++ lib/src/json/editorsettingsjson.dart
@@ -1,0 +1,28 @@
+import 'package:tiled/src/json/chunkjson.dart';
+import 'package:tiled/src/json/exportjson.dart';
+
+class EditorsettingJson {
+  ChunkJson chunksize;
+  ExportJson export;
+
+  EditorsettingJson({this.chunksize, this.export});
+
+  EditorsettingJson.fromJson(Map<String, dynamic> json) {
+    chunksize = json['chunksize'] != null
+        ? ChunkJson.fromJson(json['chunksize'])
+        : null;
+    export =
+        json['export'] != null ? ExportJson.fromJson(json['export']) : null;
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = <String, dynamic>{};
+    if (chunksize != null) {
+      data['chunksize'] = chunksize.toJson();
+    }
+    if (export != null) {
+      data['export'] = export.toJson();
+    }
+    return data;
+  }
+}
