+++ lib/src/json/tile_map_json_parser.dart
@@ -1,0 +1,13 @@
+part of tiled;
+
+class TileMapJsonParser {
+  TileMapJsonParser();
+
+  MapJson parse(String string) {
+    // TODO Check Filetype?
+    // TODO Custom parser?
+    // TODO ?.toDouble()
+    final MapJson mapJson = MapJson.fromJson(jsonDecode(string));
+    return mapJson;
+  }
+}
