+++ lib/tiled.dart
@@ -5,11 +5,12 @@ import 'dart:math';
 import 'dart:convert';
 import 'dart:typed_data';
 
+import 'package:tiled/src/json/mapjson.dart';
 import 'package:xml/xml.dart';
 import 'package:archive/archive.dart';
 
 part 'src/tile_map_parser.dart';
-part 'src/tile_map_json_parser.dart';
+part 'src/json/tile_map_json_parser.dart';
 part 'src/tile_map.dart';
 part 'src/tileset.dart';
 part 'src/image.dart';
